Adding support for connection strings, and switching to using the pool for query.
Changed the method to support either a config object with the connection info, or a connection string. Also extended with an additional argument to be able to set some of the pg.defaults.

Switched the dbExec to use the pool.query method, as opposed to obtaining a client off the pool, and then running a query on it.

Switched some string building to use interpolation.

Added an example connection string into the test file, which can be switched out for the config to test the connection string method.